requirements.txt
feedparser
google-generativeai
openai
boto3
psycopg2-binary
google-auth-oauthlib
google-api-python-client
python-dotenv
sqlalchemy